requirements.txt
# Core Deep Learning Framework
torch==2.4.1
torchvision==0.19.1

# Quantum Computing Simulation
qiskit==1.2.4
qiskit-aer==0.15.0

# Cryptography & Security
cryptography==43.0.1

# Configuration Management
PyYAML==6.0.1

# Data Processing
numpy==1.26.4

# Visualization
matplotlib==3.9.2

# Streamlit UI
streamlit==1.39.0
